# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ultra ERP API Registry",
        "description": "Global modular API documentation for the Production-grade ERP system.",
        "contact": {
            "email": "admin@erp.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Primary API Gateway"
        }
    ],
    "paths": {
...
        "/api/v1/warehouses": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "List all warehouses",
                "operationId": "772c5e5ec9e48d1d83a6cfcb0e86e255",
                "parameters": [
                    {
                        "name": "active_only",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 1,
                                                "name": "Main Warehouse",
                                                "code": "WH-MAIN",
                                                "is_active": true
                                            }
                                        ],
                                        "total": 1
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Create new warehouse",
                "operationId": "25dfc1a312f4ac6651beb13d66788553",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    },
                                    "branch_id": {
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Warehouse created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Warehouse created.",
                                    "data": {
                                        "id": 2,
                                        "name": "East Coast Hub",
                                        "code": "WH-EAST",
                                        "is_active": true
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/warehouses/{warehouse}": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get warehouse details",
                "operationId": "6a11981951cd190a14cac1dccb902036",
                "parameters": [
                    {
                        "name": "warehouse",
                        "in": "path",
                        "description": "Warehouse ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "name": "Main Warehouse",
                                        "code": "WH-MAIN",
                                        "is_active": true
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Warehouse not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Update existing warehouse",
                "operationId": "b652500ee44073d87442b7ad0855a778",
                "parameters": [
                    {
                        "name": "warehouse",
                        "in": "path",
                        "description": "Warehouse ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    },
                                    "branch_id": {
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Warehouse updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Warehouse updated.",
                                    "data": {
                                        "id": 1,
                                        "name": "Global HQ Warehouse",
                                        "code": "WH-MAIN",
                                        "is_active": true
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Warehouse not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Delete warehouse",
                "operationId": "fc1d6e0b34e0670bf5e9082b104addda",
                "parameters": [
                    {
                        "name": "warehouse",
                        "in": "path",
                        "description": "Warehouse ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Warehouse deleted successfully"
                    },
                    "404": {
                        "description": "Warehouse not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in 'Bearer {token}' format",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}